requirements.txt
GitPython
requests
schedule
